By default be verbose rather than quiet.
As discussed with Federico, that has to be an user friendly piece
of software where the user is possibly interested in having snap-
shots of what tests the tool is performing.  Hence we have chosen
to default to verbose, and make it possible for the user to shut
neubot down via -q command line option (which is similar to e.g.
cvs behavior.)  Of course, we will also start the background Neu-
bot daemon with -q, so that it does not generate too many logs in
the syslog in that case -- when the user is certainly not curious
about what the tool is doing.  Another piece of change is that we
use -q in the Makefile when we create the default configuration
file in /etc -- we don't need to annoy the user with too many me-
ssages at that point.