Revert "bin/unix/neubot: create $HOME/.neubot/ if non-existent"
This reverts commit e77fe8fb5219894350b0b197b8eef6c877f1c70a.
Probably it was not a very good idea to touch the user directory,
given the fact that one might want to run neubot in place.  So,
the correct policy re $HOME/.neubot/ seems to me that if one wh-
ish to exploit the local configuration directory, she certainly
might create that directory as well.