http/utils.py: more flexible prettyprint approach
This new approach let the caller pass the function pointer to be
invoked (in pure functional style) rather than the file to inv-
oke write() method of.  That new approach is more flexible beca-
use e.g. the logger does not export a file like interface.

In the long run, we might want to convert the old code to the new
interface and to delete the old approach.