Implement client- and server-side rendez-vous
That is implemented as follows.  There is a module that deals with
application/json message bodies, dealing with marshalling and unm-
arshalling, as well as with Unicode issues (as long as we stay wi-
th Python 2.)  We have defined two messages so far.  The former is
called 'clientinfo' and it is the one attached to the client POST
request, that bears information on the client version, on the kind
of test the client could perform behaving as a TestServer, on the
set of tests it whishes to perform.  The latter is called 'todoli-
st' and contains the server response, which includes both someth-
ing to be done and information on the latest neubot version (if
needed.)

Then there are two modules: clients and servers.  They respective-
ly implements server-side and client-side behavior -- we have ch-
osen to use the plural here because we acknowledge we might want
to implement many of such behaviors in the future.  For now, each
module implements just a stub -- the client sends a pretty-random
request, the server composes a fixed response and sends it back.
The goal of this setup is that of testing the 'control plan' of
Neubot before we actually implement the real tests and the data-
base module -- so that we can get hands-on as early as possible.

In addition we need to export the new modules so that they are
seen via 'import neubot'.  While for servers and clients we have
chosen not to export the clients (servers) part -- e.g. you can
access neubot.servers.simpleserver as neubot.simpleserver -- we
believe that the code would be much more readable if we retain
the module name for rendezvous -- e.g. neubot.rendezvous.todolist
is a more transparent name, for the casual reader, than just
neubot.todolist.

In the end, we have integrated the above code with our startup
mechanism, so that the default command, _neubot, runs the client
side, and the new command, coordinator, runs the server side.