network/pollers.py: very robust exception handling
Rationale: when Neubot encounters an unhandled exception, we do not
want that it exit()s immediatly, because we might loose information
in such case.  Instead we have issued the following policy:

    When a connection causes an unhandled exception, the poller will
    catch that, and will also close that connection.  That is a rea-
    sonable behavior from the poller point of view: it is obvious that
    such connection cannot continue at that point.  However, it is
    not obvious whether the Neubot at whole should be stopped at that
    point.

Of course, this policy conflicts with Python default policy to
map signals into exceptions (such as SIGINT that becomes an Ex-
ception.)  That could be solved overriding the default Python
behavior, or creating special-case code for such an Exception.
We will deal with this minor issue with a future patch.

For now, being in an heavy testing phase, I feel like it is
better to come out with a very pragmatic approach, so that we
can deal with exceptional behavior printing useful messages
on the system logger.  In the future, we might want to issue
something that is more fine grained on the connections party,
but I definitly think that it is acceptable, for now, if we
just close the connection whenever we encounter problems.