Provide simple watchdog mechanism to avoid getting stuck
This patch implements a simple watchdog mechanism that makes it
possible to stop an HTTP transaction that is taking too much time
to complete.  While we acknowledge that this solution is raw and
probably should be replaced by something more smart in the near
future, we have decided to roll it out because we found that some
Neubot instances sometimes hang forever in the Windows operating
system.  So, this is a quick and (a bit) dirty approach that tries
to address the problem.

More in detail, this commit makes the following changes:

  neubot/http/protocols.py

    In the constructor, register with the poller a periodically
    invoked function, and, also, save the current time.

    In the periodic function, make sure not more than TIMEOUT
    seconds have passed since this protocol instance was created.
    Otherwise, register the close function to be invoked at the
    beginning of the next poller loop.

    When closing the protocol, unregister the periodic function.

  neubot/network/pollers.py

    Make it possible to register and unregister periodic functions.

    Amend the mechanism to register initializers, modifying its name
    to reflect the now-broader scope: We employ it not just for the
    initializers, but also for all the functions that must be invok-
    ed once.  We will also retain the old function name for a while,
    together with a deprecation warning, for the code to be updated
    to the new model.

    In the poller loop

      Amend the loop to employ self.funcs rather than self.inits, as
      we changed the name, as discussed above.

      Consider self.periodic in the rule that tells whether we can
      break the loop because there is nothing to do.

      Amend the periodic invocation part, because now periodic is a
      set, while it used to be a function.

Note that it's not feasible to close the protocol in its periodic()
method, because that will change the size of the periodic set of the
poller, which we are iterating over, in turn causing a runtime exc-
eption.  For this reason, in periodic() we register self.close() to
be invoked at a later time.