Makefile: _install: Must not compile Python sources
Rationale: That target is employed by the debian package creation
process, which should not include binaries, because we provide a
.deb for all architectures.  Instead, it's the task of the postinst
script to byte-compile the sources in /usr/share/neubot.

On the other hand, we need certainly to byte-compile the sources
with the simple install target of the Makefile.

So, the solution is to move the instruction to byte-compile from
the _install to the install target.