http/protocols.py: Let the application decide whether we have body
Add the protocol the .donthavebody() method that the application might
invoke when it has knowledge that the server is not going to send a body
attached to the request.  This happens, for example, when the application
sent an HEAD request to the server.

With this amendment in place we can also measure the time required to get
the response to an HEAD requets, which approximates by excess the latency
over the communication channel.