neubot/measure.py: Measure 'time to transfer' instead of speed
Simplify a bit the model used to take measurement.  We might make it
more complex again later, however, at this point of the develop process
I prefer to roll out a simple solution that just works.

More in detail, this procedure measures the time that spans between
a moment before we start sending (receiving) the request and a moment
after we finished receiving (sending) the response.