neubot.debug: Trace functions "call" and "return"
This patch implements a profile function for sys.setprofile() that
traces "call" and "return" events.  To enable call tracing one should
do::

  sys.setprofile(neubot.debug.trace)

The trace function will print a message on the standard error when a
function is called or returns.  The message is indented depending on
the current depth of the call stack.  One example of output might
look like this one::

  call neubot/net/connectors.py:121:connect()
      call neubot/net/connectors.py:43:__init__()
          call neubot/net/connectors.py:61:_connect()
              [...]
          return neubot/net/connectors.py:85:_connect()
      return neubot/net/connectors.py:56:__init__()
  return neubot/net/connectors.py:124:connect()

This feature helps to understand the program behavior and might be
very helpful to spot places where objects are not correctly destro-
yed.  Indeed, if an object defines the __del__() method and we do
not see this method invoked where we expected to, this means there
are some extra references preventing the object's destruction.

This patch also modifies bin/unix/neubot and now the script checks
for the environment variable NEUBOT_TRACE.  If this variable exists
and has a "not false" value, the script invokes sys.setprofile()
and enables call tracing.