neubot.http: Use neubot.net instead of neubot.network
This patch modifies neubot.http for neubot.net.

The most obvious part of the changes is in acceptor and connector because
they are very close to neubot.net.listen() and neubot.net.connect().

Another simple set of changes is the one in collect.py, container.py,
measure.py, negotiate.py, rendezvous.py, getter.py and publish.py where
basically we needed to add an explicit recvmessage() where before it
was implied--indeed one of the most relevant changes is that now the
adaptor does not read from the underlying socket unless explicitly asked
to do that.

We have applied the following changes to the adaptor: (i) do not read
the socket unless explicitly asked to do that, (ii) modify the code that
actually sends and recvs data to work with neubot.net rather than with
neubot.network.  We also have fixed the closing() function to make sure
that there are not circular dependencies.

The protocol has been modified to add the .recvmessage() member function,
that asks the adaptor to start receiving from the socket--because we have
amended the adaptor as explained above.  And we have also fixed closing()
to avoid circular dependencies.