neubot.http: Add message-oriented API
This patch wraps the existing neubot.http code with a new message-oriented
API--the general idea beyond this API is that you declare what you want to
do, like sending an HTTP message, and the module machinery then performs
that action for you.  You can register callbacks to be invoked on success,
and on failure, as well as to intercept some low level events--such as
that some bytes have been sent.

This API exports four functions: compose(), reply(), send(), and recv().

Compose() is the general purpose function to create a "message", and a
message is an object that contains information on what the module should
do when the message itself is passed to send() or recv().  There are
two different type of messages: the connected and the unconnected ones.
A message is connected if it holds a reference to a connection, and is
not connected otherwise.  The behavior of send(), and recv() is very
different from when the message is connected to when it is not connected.

It is possible to compose a message "in reply to" another message--in this
case the new message will inherit the reference to the connection from
the older one.  Reply() is just a convenience function that makes it more
explicit that a message is "in reply to" another one.

Send() attempts to send the message passed as its first argument, and
will use the message connection, if the message is connected.  Otherwise,
if the message is unconnected, send() will connect to the remote host,
extracting address and port information from the uri, and then will send
the message over such new connection.

Recv() is more complex than send--if the message passed as its first
argument is connected, recv() attempts to receive a response to such
message.  Otherwise, the message is interpreted as a request to listen
and accept connections incoming on the address and port specified in
the message itself.  Once the connection is accepted, recv() will read
and properly dispatch the incoming HTTP message.

[Please note that, while the API is quite stable, the implementation
 isn't, and that happens, in particular, because this API started as
 a wrapper around the old implementation, in order for this new API
 and for the old code to be available at the same time.  This is very
 handy, but caused the code to become a bit clumsy--so we need to
 follow-up with some refinements as soon as the old code is no longer
 directly employed by the Neubot classes.]