neubot.whitelist: Temporarily disable white-listing
I found that the current white-listing scheme prevents some legitimate
measurements to be performed.  This happens when there is more than one
client running on the same personal computer.

This patch is a simple and radical solution to the problem, and it just
disables the white-listing mechanism.  I will follow-up with a better
solution soon.