neubot.net: streams.py: Deal with "continuous" I/O operations
The "continuous" I/O operation case is when you are streaming some data,
either reading or writing.  In this case it is possible to optimize the
streams code a little bit (tests suggest that we can gain 8% or more.)

While optimizing a bit for "continuous" I/O, this patch also stops using
the compat flag, because (i) the code that required it now does not require
it anymore, and (ii) without the extra complexity of the compat flag it
is simpler to implement the optimization below.

More in detail: In case of successful read, the existing code unsets
readable /before/ invoking the "successful read" callback--and, in a
similar fashion, unsets writable before invoking the "successful write"
callback.  This strategy is conservative, because the callback might
change the state of the stream, invoking stream.recv() or .send().

Instead, the optimization moves the code that decides whether to unset
readable /after/ the callback invocation, and so, if the callback invokes
recv() again, we avoid an unset_readable() call, as follows::

  # Before                                # After
  _do_recv()                              _do_recv()
    sorecv() -> SUCCESS                     sorecv() -> SUCCESS
    recv_pending = False                    recv_pending = False
    UNSET_READABLE                          # We gain here
    notify()                                notify()
      recv()                                  recv()
        recv_pending = True                     recv_pending = True
        SET_READABLE                            # Moved below
        _do_recv()                              _do_recv()
          <<already receiving>>                   <<already receiving>>
        _do_recv()                              _do_recv()
      recv()                                  recv()
    notify()                                notify()
    #                                       SET_READABLE  # XXX
  _do_recv()                              _do_recv()

We exploit the fact that we set recv_pending to False before invoking
notify(), and recv() sets recv_pending to True.  Therefore, if it is
True after notify(), this means that recv() has been invoked.  So, we
gain something because we avoid an unset_readable call.  However, it
is possible to further improve the code, because it invokes set reada-
ble no matter whether the stream is already readable or not (we do
that because of "optimistic" I/O--and we can improve the code setting
readable iff we're not already in the readable state.)

There is also an extra complexity--after notify() it is possible for
recv to be blocked, for example because notify() invokes send() and
sosend() returns WANT_READ, as follows::

  _do_recv()
    sorecv() -> SUCCESS
    notify()
      send()
        _do_send()
          sosend() -> WANT_READ
          <<block recv>>
        _do_send()
      send()
    notify()
    # XXX be careful here
  _do_recv()

The solution is to be very careful after notify() returns--just do not
touch the stream status if recv is blocked.