neubot.http: Add support for passive close()
The original neubot.http.adaptor was always checking the underlying
socket for readability, and, for this reason, we were receiving EOFs
promptly, when the peer closed the connection.  In recent releases
we have changed that, and the result was that we were missing those
EOFs.

This patch adds the passiveclose() method to neubot.http.protocol,
and this method adjusts the watchdog timeout to one second (so that
we close the connection if the client does not close it after one
second,) and tries to read a message (in order to get notification
of EOF.)