neubot.whitelist: Allow more than one peer per IP address
This patch provides a quick and dirty solution to the problem pointed out
by 4de20f564bb... "Temporarily disable white-listing".  In that patch we
disabled white-listing because we found that it assumed just one Neubot
per IP address, and that prevented many legitimate use-cases, such as when
you are testing a new version of Neubot.

Here we provide an initial solution to this problem, keeping a counter that
counts the number of negotiated tests for each IP address.  This solution is
not particularly brilliant, as noted in a comment to the new code, and should
only be seen as a temporarily fix.

What we actually need is to enable measurements from peers that provide us
with a valid measurement identifier (for example via cookies for HTTP or
via infohash for BitTorrent) because it's the identifier what we actually
negotiate.