neubot.net: streams.py: Block send(), recv(), close() if isclosing
This patch blocks all the public functions of a stream, send, recv,
and close, when we enter in "closing" mode--i.e. when the poller has
invoked our closing() method.  This avoids either the user invoking
send or recv, and therefore preventing stream destruction because the
stream might be set readable or writable again, or infinite recursion
where closing() causes close() to be invoked, that in turn causes again
closing() to be invoked, and so forth [This behavior was spotted when
testing the stream optimization patch, and, indeed, has actually nothing
to do with the optimization itself.  It is completely another beast.]