neubot.ui: RESTful API for user interfaces attachment
The idea beyond this patch is that there is a certain number of
state variables in the Neubot code that user interfaces should be
able to access and/or modify.  And we want to export these vari-
ables via a RESTful interface for many different user interfaces
to be able to control Neubot.

We begin with exporting a simple fine-grained text/plain API and the
option to return results in HTML format--but, of course, the future
direction is towards coarse-grained JSONp access, and this should enable
users to control Neubot from HTML5-enabled browsers.

This patch also includes a very simple and powerful text-only test
client, that could be used to enable control from command line as
well.