neubot.http: Send small messages using a single send()
This is important because (i) if we successfully pass the message to
the socket layer using one single send() we avoid setting writable the
socket descriptor, and so (a) we avoid employing more than one poller
loop to send a message and (b) select() will run faster (since it is
O(n) where n is the number of sockets we're polling for I/O events,)
and (ii) the receiver might receive the message in a single packet
(for very small messages) and so it should be able to process such http
message in a single poller loop.

While at it, remove some quirky bits from the adaptor implementation.