neubot.config: Add 'force' variable to configuration
When this variable is True neubot will do "unreasonable" (in the opinion
of the developer) things, such as performing a transmission test when
the rendez-vous takes too much time to complete.  The underlying idea of
this patch is that the user knows what to do and should be able to work
around our (possibly broken) assumptions.