neubot.notify: Publish/subscribe mechanism
We will implement Comet using this mechanism--consider for example the
case when we want to follow the evolution of the program state (which is
what the icon in the notification area does:)  the delayed response is
either generated when the program state changes, or when a certain time-
out expires.

When we delay HTTP responses, the underlying connection might be closed
by the remote host.  In this case the stream and the adaptor are correct-
ly destroyed, but the protocol isn't because the request message holds
a reference to it, and, in turn, this module keeps a reference to the
request message (and this reference is needed to delay the response.)
So, if the connection is closed, the protocol will try to send the delay-
ed response over a None adaptor.  This raises an exception and the Ne-
ubot server will eventually crash--and, for this reason, we surround
with try..except the func() invocation [see the patch below] and this
will prevent the crash.  Of course, this is not a solution to the problem
but just a patch that keeps the crash away.  We will roll-out a better
solution in the next release cycle.  And, by the way, we name this sort
of problem "Comet after close".