neubot.net: poller: Remove unneeded catch-all clauses
The stream code catches socket and SSL errors when reading and
writing--and so we don't need anymore the catch-all clauses in
poller's readable() and writable() functions (with the code in
neubot/network/connections.py we needed to do that because it
re-raised exceptions.)  Yes, these catch-all clauses have the
benefit that they catch programming errors, such as when one has
mispelled a function name.  But this is something that should
live at the highest layer of abstraction, and not in the network
layer.