neubot.ui: Export state change via comet
The general idea is that one can GET either '/state' or '/state/change',
and, in the former case, the response is the program state in the moment
of the request, and, in the latter case, the response is delayed until
the program state changes _or_ a given timeout expires.

To track the evolution of the program state, the icon in the notification
area should therefore issue a 'GET /state' and, once it knows the current
state of Neubot, keep track of further changes with many subsequent (and
possibly delayed) 'GET /state/change' responses.