neubot.net: streams: Optimize readable and writable management
This patch adds a tiny layer of abstraction to the Stream class, and
this layer makes sure that we invoke poller's set/unset readable or
writable only if needed.  In other words, if we already registered for
readability with the poller we will not try to register again, and,
similarly, if we're not registered we will not try to unregister.