neubot.net: pollers: Modify the task scheduler to support periodic()
This patch is detached from the one that implements the task scheduler
because it is needed as an interim measure to emulate periodic(), but
we plan to remove it as soon as the existing code employs sched() and
unsched() rather than register and unregister periodic().  And so we
keep the patch detached because we will need to revert it in the futu-
re.

Note that periodic functions are different from ordinary tasks because
they receive the current time as their only argument.