neubot.utils: Add support for keyword arguments
For functions that receive too many arguments, it would be nice to
exploit keyword arguments.  But, when you employ keyword arguments,
you need to check whether an argument was supplied or not, and, if
not, you need to assign a default value, as follows:

  ...
  if kwargs.has_key("foo"):
    foo = kwargs["foo"]
  else:
    foo = DEFAULT_FOO
  ...

This is a bit annoying--and here it comes fixkwargs().  This is a
function that preprocess kwargs, and sets the default value for
arguments that were not supplied.  It roughly works as follows:

FOOBARARGS = {
  ...
  "bar" : DEFAULT_BAR,
  "foo" : DEFAULT_FOO,
  ...
}

def foobar(**kwargs):
  kwargs = fixkwargs(kwargs, FOOBARARGS)
  foo = kwargs["foo"]
  ...

After fixkwargs() kwargs contains either the supplied value or
the default value, for each valid argument.  Other arguments are
ignored.

This patch is the first step in the effort of simplifying functions
that receive too much arguments--more patches will follow-up.