testing/wget.py: Automatically prepend "." to sys.path
This patch is part of the ongoing effort to make all the standalone
modules easily invokable from the command line (we define "standalone
module" either a testing module or a neubot module that one might
want to invoke from the command line.)

To avoid setting PYTHONPATH at hand each time we want to invoke a
standalone module, we need to /prepend/ "." to sys.path.  It is very
important to prepend and not to append, because we want the standalone
module to employ the Neubot Python Library in ".", and not some old
version of NPL that might exist in the path.