neubot.net: pollers: Remove .closing() from Pollable
In Neubot 0.1.7 I have removed catch-all clauses from the poller code,
because I wanted to move them at higher abstraction layers.  Before this
patch, the .closing() callback of Pollable was invoked in case of (a)
caught exception, or (b) when a stream was closed.  So, without the
catch-all clauses, the .closing() callback is just an elaborate way to
invoke the stream destructor, as follows::

  stream.close(self)
    poller.close(stream)
      stream.closing()
        <<destructor code>>

This patch removes the .closing() callback and moves the stream destr-
uctor code in .close().