neubot.http: protocols: Do not lower timeout for passiveclose
I found that the Neubot client was not correctly performing transmission
tests when running over an ADSL connection, and it seems the responsible
is passiveclose.  In particular, one second of timeout for passiveclose
might not be enough when sending to an ADSL connection (while it is enough
when testing over loopback, and, indeed, I did not noticed any problem
when I were at NEXA.)  I am not sure why one second is not enough--is it
possible that close() is invoked by passiveclose too early, when the TCP
sender has not completely flushed its send buffer?  For now, just keep
the huge original watchdog timeout value because it does not disrupt the
tests.