neubot.net: pollers: Remove 'Nothing to do' warning and sleep
When we exit from the poller because we have nothing to do it is
pointless to print a warning saying that.  And it is annoying to
sleep in this case, because this delayed exiting from the core
loop.  While this behavior made sense when the poller employed the
one-shot+periodic funcs model, it has become useless and even a
bit annoying now with the task scheduler.