neubot.net: pollers: Increase interval between check_timeout()
Recently we modified pollers loop() and now the loop continues as long
as there are either readable or writable sockets.  This amendment makes
it possible to raise the interval between each check for timed out I/O
operations.  This was not pratical before we modified the loop, as shown
by the following example:

1. There is a readable socket.

2. Check_timeout() runs and schedules its next invocation in N secs.

3. The socket is closed for some reason.

4. The loop continues until N seconds are expired BECAUSE there is
   a scheduled function.

5. Check_timeout() runs but does not schedule itself because it notices
   that there are no readable or writable sockets.

6. Finally we break out of the loop.

So, we were waiting, on the average, N/2 more seconds in the loop, and
this was acceptable with N=1 but it was clearly not acceptable with, say,
N=10.

Now that the loop continues as long as there are readable or writable
sockets, it is pratical to raise the timeout, because the loop WILL NOT
continue, when there is JUST a scheduled function.

And so this patch increases the interval between each check.