neubot.http: Add new sub-module: handlers.py
This module replaces the adaptor, and tries to be more efficient than
the adaptor, bufferizing as few incoming data as possible.  It tries
also to be convenient, and, thus, to send() an HTTP message a sender
just needs to register some optional callbacks.  Finally, it is cleaner,
because it provides a better separation of the MECHANISM to read an
HTTP message and the POLICY.  This class implements the mechanism, and
the policy decision is up to the class controller.

A companion patch (which we will roll-out in the next release cycle)
implements a reasonable policy.  And so, these two pieces together are
a solid framework to implement HTTP (and probably we will base more
than one APIs on top of them).