http.utils: Add rules to decide the nextstate after headers
This patch adds the rules to decide the next state of the handler
after we got headers.  The rules are taken directly from the RFC
that defines HTTP/1.1 and they depend on whether we are receiving
a request or a response.  These rules are not implemented in the
handler because:

1. The scope of the handler is the syntax and not to decide the
   body type, which is semantic stuff because it requires header
   analysis.

2. We might want to specialize these rules--for example in the API
   we might not accept chunked bodies or huge bounded bodies, and
   this makes sense because the point of the API is to exchange
   *small* messages to coordinate the Neubots.