neubot.net: pollers: Make sure unsched() builds a 3-tuple
Sched() builds a 3-tuple (delta, func, periodic) and so unsched()
must build the corresponding 3-tuple when searching for an event
that is in self.added.  If we don't build a 3-tuple in unsched(),
we will fail to unschedule the event, because the comparison of
a 3-tuple and a 2-tuple is always False.

While in many cases we unschedule events that are already registered,
there are also some cases where we unregister events that are not yet
registered.  For example, we might send the request using the optimistic
send feature, register the stream of readability, receive and process
the headers, register the event, find out that there is no body, close
the connection, break out of the poller loop, and, finally, unregister
the event.  In this case, the event is added to self.added and never
moved to self.tasks, because we employ just one poller loop to receive
the response--and so we do not invoke update_tasks().