net.pollers: Explain why we break loop() regardless of tasks
We break loop() regardless of tasks because: (i) neubot is not ready
to do everything inside loop(), and the code often assumes that loop()
will break as soon as I/O is complete; (ii) there are some tasks that
re-schedule self forever, like the one of neubot.notify.  So, either
we break out of the loop regardless of tasks, or we will loop forever
due to some self-scheduling task, and this might not be exactly what
we want.