http.messages: Move some to-be-removed fields out of Message
While some parts of api.py, such as compose(), are cool, some
others are too abstract and refining the implementation seems
a bit too complex--probably this is a symptom of bad design and
so I should not insist with it, and I should roll-out something
simpler instead.

So, let's start moving out of Message the bits that belong to
api.py and that will be removed together with it.

While at it, use AF_UNSPEC instead of 0.