http: Move compose_() from api to messages
Move compose_() from api.py to messages.py, and, while at it, rename
it compose().  Note that the renaming does not generate a conflict
because messages.py exports just 'message'.  And so, the compose() that
neubot.http exports is api.compose() and not messages.compose().  Also,
api.py access messages.compose() as compose_() to avoid naming problems.

Now we can confess it--the whole compose_() stuff was part of an effort
to create a better compose() in http.messages.  In the long run api.py
will be removed, and messages.compose() is going to become *the* compose()
of Neubot.