http.clients: New module that implements HTTP client
While api.py contains interesting bits it's being a bit difficult
to finalize its implementation, perhaps because I made the mistake
of trying to design something "cool".  Ok, probably it's time to
review once again the keep-it-simple lesson rolling-out a piece of
code that "just works".

This patch adds the try neubot/http/clients.py that contains Simple-
Client and Client.  The former is a simple wrapper around Handler
that adds the point of view of a client.  The latter extends Client
with: (i) code to measure performances; (ii) convenience functions
for the measurements.  The module itself could be invoked directly
from the command line to perform upload, download, and latency tests.