http.servers: New module that implements HTTP server
While api.py contains interesting bits it's being a bit difficult
to finalize its implementation, perhaps because I made the mistake
of trying to design something "cool".  Ok, probably it's time to
review once again the keep-it-simple lesson rolling-out a piece of
code that "just works".

This patch adds neubot/http/servers.py that contains SimpleConnection,
Connection, and Server.  SimpleConnection wraps the Handler and provides
the point of view of a server.  Connection extends SimpleConnection with
code to measure performances and convenience functions for the measure-
ments.  Server listens for incoming connections.

The module itself could be invoked from command line to serve files,
using the current working directory as DocumentRoot.