http.handlers: Improve algorithm to split first-line
This patch improves the algorithm that splits the first line of an
HTTP message.  The new algorithm employs strip() to remove trailing
newlines from the line, and so, when we tokenize the line, the last
token is not CRLF terminated.  In addition, set the maximum number
of tokens to 3, and so we don't fail to parse responses whose reason
includes spaces.