neubot.net: pollers: Measure average upload/dload speed
In this patch we need to copy formatbytes() from http.clients because
it was not possible to import easily that function from this module.
This probably means that it's better to keep formatbytes() here in the
net module rather than in http.  In the next release, or so, we will
replace http.clients formatbytes() with an import statement that pulls
formatbytes() from net.pollers.  And, in a couple of releases, we will
ensure that all modules load formatbytes() from net.pollers and not
from http.clients.