http.clients: Support for download using two connections
This patch modifies http.clients to add support for downloading
a resource using two connections.  When the round trip delay
is high but there's a lot of available bandwidth, much time is
wasted waiting for ACKs and a great amount of available bandwidth
is not employed.  And so, using two connections, it's possible
to speed the download up, because, when a transmitter is idle
waiting for the ACKs, the other might be able to send over the
socket.