http.clients: Avoid race between _do_send() and closing()
When we connect and the remote server suddenly closes the connection,
there is a race between _do_send() and closing().  The problem is that
send() invokes _do_send() after handler.start_receiving(), and, if the
connection is closed, the latter invokes closing(), that clears the
reference to the handler.  So, when _do_send() runs, it operates on a
None handler, and so Python raises an exception.

The fix is simple--before invoking _do_send(), make sure that the
handler is still not None.