http.servers: Avoid race between constructors and handler
Let's assume we employ Connection rather than SimpleConnection. Well,
in this case it's wrong to start_receiving() in SimpleConnection's
constructor, because, if there is already something to read(), we will
start processing incoming data BEFORE Connection.__init__() completed.
And so, the Connection object is not complete, and certain fields are
missing, and Python will raise an Exception.

There are two possible fixes for this issue.  The former is to re-
write Connection.__init__() to construct its fields before invoking
the parent's constructor.  The latter is to avoid start_receiving()
in SimpleConnection.__init__() and to create, instead, an ad-hoc
function, start_receiving(), that should be invoked when the chain
of constructors has returned.

We choose the latter, because it is more robust: We don't need to
keep in mind that the order matters for SimpleConnections childs'
constructors.