ui: It is wrong to append "." to sys.path
If we append, an already installed copy of Neubot might take
precedence.  And this is not what we want when we run unit
testing code, because, instead, we want to pick the freshest
copy of Neubot, the one in the current working directory.  And
so it's better to use insert(0, ".").