http.servers: Detect and handle comet-after-close
We define comet-after-close the case when the server has a
delayed comet response and the client already closed the
connection.  When the connection is closed, the net layer
notifies the Handler, and the Handler, in turn, notifies
the Connection, and the Connection clears its reference to
the handler.  But the Connection itself is not destroyed
because there is not a mean to tell notify.py that such
Connection has been closed [of course, we can implement it,
but that is beyond the scope of this patch].  So, when it
is time to fire the delayed response, we must be prepared
for the comet-after-close case, which means that we must
check whether self.handler is None or not.