ui: Rewritten UIServer in order to rationalize its implementation
Now the code should be more readable and easy to be modified, and we
don't need anymore to generate HTML at hand, and we have replaced all
those redundant checks to generate 405, 415, etc--we had a check for
every if..else and now we have instead an algorithm that works on a
table that represents all the existing URIs.

Among the changes, there is one that modifies the keepalive behavior
of UIServer.  Now it does not close the connection by default, and so
the client should be modified because it assumes the server implements
the old behavior.