net.pollers: Make sure that we export register_func
The function periodic() of http.protocols employs register_func()
but net.pollers does not export it.  And so the server crashes in
case of 'Watchdog timeout'--to avoid this issue we just need to
export register_func().

BTW, this seems a good place to discuss the switch from using import
to using from..import.  If we use the former and we access some non-
existent attribute of the module, then there is an Exception when the
program is running.  On the contrary, if we use the latter, there is
an ImportError when Python is compiling the program.  And so, in the
former case we might not be aware of the error because the offending
line might not be frequently touched, while, in the latter case, we
always see the error.

On the other hand the code seems more tidy with the import approach,
and the code is more readable because you name the package in front of
each function you invoke.  BUT, naming packages cause lines to be longer
and this might impact on readability.

So, probably, we can wrap up it this way--the new approach catches more
errors and makes the code more readable (because lines are longer) but
less tidy (because there are much more imports).